# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for FanFic AI
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Verify admin password and return a session token (currently equal to the password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid password
  /admin/stories:
    get:
      operationId: adminListStories
      tags: [admin]
      summary: List ALL stories (drafts + published) with metadata, requires admin token
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: All stories with stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminStoryRow"
        "401":
          description: Missing or invalid admin token
  /admin/stories/{id}:
    delete:
      operationId: adminDeleteStory
      tags: [admin]
      summary: Permanently delete a story and all its data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Missing or invalid admin token
        "404":
          description: Not found
    patch:
      operationId: adminUpdateStory
      tags: [admin]
      summary: Update a story's status (publish/unpublish) as admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateStoryBody"
      responses:
        "200":
          description: Updated story
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
        "401":
          description: Missing or invalid admin token
        "404":
          description: Not found
  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Aggregate site statistics
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stats payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Missing or invalid admin token
...
  /admin/tariffs/{tier}:
    get:
      operationId: adminGetTariff
      tags: [admin]
      summary: Get a tariff (free)
      parameters:
        - name: tier
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Tariff, content: { application/json: { schema: { $ref: "#/components/schemas/Tariff" } } } }
    put:
      operationId: adminUpdateTariff
      tags: [admin]
      summary: Update a tariff's daily limits
      parameters:
        - name: tier
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTariffBody" }
      responses:
        "200": { description: Updated tariff, content: { application/json: { schema: { $ref: "#/components/schemas/Tariff" } } } }
...
  /stories/{id}/view:
    post:
      operationId: recordStoryView
      tags: [stories]
      summary: Record a story view (analytics)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StoryViewBody" }
      responses:
        "204": { description: Recorded }
  /stories/{id}/analytics:
    get:
      operationId: getStoryAnalytics
      tags: [stories]
      summary: Author analytics for a story (last 30 days)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: authorName
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Analytics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StoryAnalytics" }
...
components:
  schemas:
...
    Story:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        genre:
          type: string
        artStyle:
          type: string
        lengthSetting:
          type: string
          enum: [short, medium, long]
        seedPrompt:
          type: ["string", "null"]
        fullText:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        characters:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published]
        authorName:
          type: string
        coAuthors:
          type: array
          items:
            type: string
          default: []
        coverImageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        likeCount:
          type: integer
        commentCount:
          type: integer
      required:
        - id
        - title
        - genre
        - artStyle
        - lengthSetting
        - status
        - coAuthors
        - authorName
        - createdAt
        - updatedAt
        - likeCount
        - commentCount
...
    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
          minLength: 1
      required: [password]
    AdminLoginResponse:
      type: object
      properties:
        token:
          type: string
      required: [token]
    AdminStoryRow:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        authorName:
          type: string
        status:
          type: string
          enum: [draft, published]
        genre:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        likeCount:
          type: integer
        illustrationCount:
          type: integer
      required: [id, title, authorName, status, genre, createdAt, updatedAt, likeCount, illustrationCount]
    AdminUpdateStoryBody:
      type: object
      properties:
        status:
          type: string
          enum: [draft, published]
        title:
          type: string
      additionalProperties: false
    AdminStats:
      type: object
      properties:
        totalStories:
          type: integer
        publishedStories:
          type: integer
        draftStories:
          type: integer
        totalIllustrations:
          type: integer
        totalLikes:
          type: integer
        totalAuthors:
          type: integer
      required: [totalStories, publishedStories, draftStories, totalIllustrations, totalLikes, totalAuthors]
...
    Tariff:
      type: object
      properties:
        tier: { type: string }
        storyDailyLimit: { type: integer }
        illustrationDailyLimit: { type: integer }
        updatedAt: { type: string }
      required: [tier, storyDailyLimit, illustrationDailyLimit, updatedAt]
    UpdateTariffBody:
      type: object
      properties:
        storyDailyLimit: { type: integer, minimum: 0, maximum: 1000 }
        illustrationDailyLimit: { type: integer, minimum: 0, maximum: 5000 }
      additionalProperties: false
...
    StoryAnalytics:
      type: object
      properties:
        storyId: { type: integer }
        totalViews: { type: integer }
        totalCompleted: { type: integer }
        totalLikes: { type: integer }
        totalComments: { type: integer }
        daily:
          type: array
          items:
            type: object
            properties:
              day: { type: string }
              views: { type: integer }
              completed: { type: integer }
            required: [day, views, completed]
      required: [storyId, totalViews, totalCompleted, totalLikes, totalComments, daily]
    StoryViewBody:
      type: object
      properties:
        viewerName: { type: ["string", "null"] }
        completed: { type: boolean }
      additionalProperties: false